reorganize swarm passives, add tooltip arguments (unused)
