--- Common/Data/Affixes/TalismanAffixes.json.old
@@ -1,3 +1,0 @@
-{
-
-}
