+++ Common/Mechanics/SkillPassive.cs
@@ -7,7 +7,9 @@ public abstract class SkillPassive(SkillTree tree) : SkillNode(tree)
 {
 	public override string TexturePath => $"{PoTMod.ModName}/Assets/SkillPassives/" + Name;
 	public override string DisplayName => Language.GetOrRegister("Mods.PathOfTerraria.SkillPassives." + Name + ".Name", () => GetType().Name).Value;
-	public override string DisplayTooltip => Language.GetOrRegister("Mods.PathOfTerraria.SkillPassives." + Name + ".Tooltip").Value;
+	public override string DisplayTooltip => Language.GetOrRegister("Mods.PathOfTerraria.SkillPassives." + Name + ".Tooltip").Format(TooltipArguments);
+
+	public virtual object[] TooltipArguments { get; }
 
 	/// <summary> The effects of this skill passive. </summary>
 	public virtual void PassiveEffects(ref SkillBuff buff) { }
