--- Content/SkillPassives/SwarmPassives/CarnivorousLarvae.cs
@@ -1,6 +1,0 @@
-﻿using PathOfTerraria.Common.Mechanics;
-using PathOfTerraria.Common.Systems.Skills;
-
-namespace PathOfTerraria.Content.SkillPassives.SwarmPassives;
-
-internal class CarnivorousLarvae(SkillTree tree) : SkillPassive(tree);
