--- Content/SkillPassives/SwarmPassives/ColdBlooded.cs
@@ -1,6 +1,0 @@
-﻿using PathOfTerraria.Common.Mechanics;
-using PathOfTerraria.Common.Systems.Skills;
-
-namespace PathOfTerraria.Content.SkillPassives.SwarmPassives;
-
-internal class ColdBlooded(SkillTree tree) : SkillPassive(tree);
