--- Content/SkillPassives/SwarmPassives/IceVenom.cs
@@ -1,6 +1,0 @@
-﻿using PathOfTerraria.Common.Mechanics;
-using PathOfTerraria.Common.Systems.Skills;
-
-namespace PathOfTerraria.Content.SkillPassives.SwarmPassives;
-
-internal class IceVenom(SkillTree tree) : SkillPassive(tree);
